requirements.txt
ai-engine-sdk==0.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
anyio==4.4.0
apispec==6.6.1
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bech32==1.2.0
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cosmpy==0.9.2
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
ecdsa==0.18.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.1
grpcio==1.64.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
newsapi-python==0.2.7
numpy==1.26.4
openai==0.28.0
opencage==2.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycryptodome==3.20.0
pydantic==1.10.16
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.36.0
tenacity==8.4.2
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uagents==0.12.2
urllib3==2.2.1
uvicorn==0.20.0
vaderSentiment==3.3.2
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
yarg==0.1.9
yarl==1.9.4
yfinance==0.2.40
